reverse mutation reproduction looping order
I was seeing this weird structure in the population genome when mutation
was used, but I realized this is just an artifact of the fact that with
a small genome size and a small error probability in expectation fewer
than 1 bits are flipped in the reproduction process. So, since many
parents will just be cloned into the child population (and their
fitnesses remain constant), you get this behavior where over time the
clones come to dominate the population.